feat(service): add domain model and remove coupling of entity again
fix strategy and dosis respect to insert entities

+++ src/main/kotlin/io/github/kingg22/api/vacunas/panama/modules/direccion/persistence/DireccionPersistenceService.kt
@@ -59,7 +59,7 @@ interface DireccionPersistenceService {
      * @param distritoId The ID of the district.
      * @return A list of address entities.
      */
-    suspend fun findDireccionByDescripcionAndDistritoId(descripcion: String, distritoId: Int): List<Direccion>
+    suspend fun findDireccionByDescripcionAndDistritoId(descripcion: String, distritoId: Short): List<Direccion>
 
     /**
      * Finds addresses by description and district name.
